small tactic_page structure change
